+++ src/export/tex/gather.ts
@@ -1,6 +1,6 @@
 import path from 'path';
 import { Blocks, VersionId, KINDS, convertToBlockId } from '@curvenote/blocks';
-import { DocumentModel } from 'export';
+import { DocumentModel } from '../model';
 import { Block, Version } from '../../models';
 import { ISession } from '../../session/types';
 import { getChildren } from '../../actions/getChildren';
@@ -53,7 +53,7 @@ export async function gatherAndWriteArticleContent(
     templateOptions,
     {
       path: opts.texIsIntermediate ?? false ? '.' : '..', // jtex path is always relative to the content file
-      filename: path.basename(opts.filename),
+      filename: opts.filename,
       copy_images: true,
       single_file: false,
     },
@@ -61,9 +61,7 @@ export async function gatherAndWriteArticleContent(
     Object.keys(article.references).length > 0 ? 'main.bib' : null,
   );
 
-  const filename = opts.template
-    ? path.join(buildPath, path.basename(opts.filename))
-    : opts.filename;
+  const filename = opts.multiple ? path.join(buildPath, 'chapters', opts.filename) : opts.filename;
   session.log.debug(`Writing main body of content to ${filename}`);
   session.log.debug(`Document has ${article.children.length} child blocks`);
   writeBlocksToFile(
