+++ src/export/tex/types.ts
@@ -1,7 +1,9 @@
 export interface TexExportOptions {
   filename: string;
+  multiple?: boolean;
   images?: string;
   template?: string;
+  templatePath?: string;
   options?: string;
   useBuildFolder?: boolean;
   texIsIntermediate?: boolean;
